// apps/web-svelte/src/lib/api/generated/model/reviewManagedBookRequestInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ManagedBookRenditionChoice } from './managedBookRenditionChoice';
import type { ReviewedRequestCommitRequest } from './reviewedRequestCommitRequest';

export interface ReviewManagedBookRequestInput {
  /** @nullable */
  entityId: string | null;
  request: null | ReviewedRequestCommitRequest;
  renditions: ManagedBookRenditionChoice[];
}